refactor(broker): split order_evaluator.rs into directory layout (C34 of PR-3)
Pure mechanical move. order_evaluator.rs (803 lines) becomes:

- order_evaluator/mod.rs (~31 lines): module decls + pub(crate) use
  service::OrderEvaluator + pub(crate) use types::{PreflightComplete,
  PreflightOutcome}.
- order_evaluator/error.rs (~36 lines): OrderEvaluatorErr enum +
  coded_error_impl!.
- order_evaluator/types.rs (~46 lines): PreflightOutcome enum + its
  Display impl + PreflightComplete (preflight-completion messages
  exchanged with the per-chain OrderPricer).
- order_evaluator/service.rs (~762 lines): OrderEvaluator struct + impl,
  BrokerService impl, the two module constants
  (MIN_CAPACITY_CHECK_INTERVAL, DEFAULT_MAX_PREFLIGHT_DURATION_SECS),
  and the original test module.

External paths crate::order_evaluator::{OrderEvaluator, PreflightComplete,
PreflightOutcome} preserved via the re-exports.

PreflightOutcome import moved into the test module — non-test code only
formats it via the Display impl on PreflightComplete.outcome, same pattern
as C33.

The canonical plan's optional order_evaluator/prioritization.rs wrapper is
omitted: it would be pure indirection over the existing
crate::prioritization::prioritize_orders_to_evaluate shared helper.
