+++ crates/broker/src/order_evaluator/error.rs
@@ -1,0 +1,36 @@
+// Copyright 2026 Boundless Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Error type for the order evaluator service.
+
+use std::sync::Arc;
+
+use thiserror::Error;
+
+use crate::coded_error_impl;
+use crate::errors::CodedError;
+
+#[derive(Error)]
+pub(crate) enum OrderEvaluatorErr {
+    #[error("{code} Config read error: {0}", code = self.code())]
+    ConfigReadErr(Arc<anyhow::Error>),
+
+    #[error("{code} Stale capacity for order {order_id} expired after {elapsed_secs}s with no completion", code = self.code())]
+    StaleCapacity { order_id: String, elapsed_secs: u64 },
+}
+
+coded_error_impl!(OrderEvaluatorErr, "OE",
+    ConfigReadErr(..)   => "001",
+    StaleCapacity { .. } => "004",
+);
